Increased the maximum possible length of the LED strip.

+++ main/led_strip/include/led_strip.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-void LED_Strip_Init(uint8_t * leds, uint8_t count);
+void LED_Strip_Init(uint8_t * leds, uint16_t count);
 void LED_Strip_Update(void);
 void LED_Strip_SetPixelColor(uint16_t pixel, uint8_t r, uint8_t g, uint8_t b);
 void LED_Strip_Rotate(bool direction);
